requirements.txt
# Core dependencies
opencv-python>=4.8.0
openai>=1.0.0
yt-dlp>=2024.1.0

# Local LLM support (optional - for fully offline mode)
ollama>=0.1.0

# Web UI
gradio>=4.0.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL driver for Supabase
pgvector>=0.2.4  # Vector extension for PostgreSQL

# Redis (Job Queue + Caching)
redis>=5.0.0
rq>=1.15.0  # Redis Queue for background jobs (optional for now)

# Embeddings (for pgvector)
sentence-transformers>=2.2.0  # For generating embeddings

# Billing (Stripe)
stripe>=7.0.0

# Email (Password Reset)
resend>=0.7.0

# Google OAuth
google-auth>=2.0.0
